Rewrite vue-mode.el
Rewrites nearly every s-expression for cleanliness, allows customization of
language-to-mode bindings, uses a more suitable data type for languages,
adds a customization group, and leaves cleaner overlay traces.

Redefines vue-mode as a subset of html-mode, in the interest of not
pulling web-mode as a dependency. Also allows explicit html-mode
specification, adds a typescript hook, and renames "default" to nil for
cleanliness' sake