Add support for indirect component section editing
`vue-mode-edit-indirect-at-point` opens the subsection at point in a separate
window, applies the appropriate major mode, and allows the user to edit that
section of the as if it were a standalone buffer.

This should greatly improve user experience in relation to auto-completion
plugins, linter plugins, and spellcheckers, which are historically difficult to
use in mmm-mode.

I feel this feature is enough justification to bump the version, so let's all
welcome vue-mode 0.3!